Use redux-logic for errors
