+++ src/logic/index.js
@@ -2,12 +2,12 @@ import linkGithubIdentity from './linkGithubIdentity';
 import logout from './logout';
 import startAccountMigration from './startAccountMigration';
 import unlinkGithubIdentity from './unlinkGithubIdentity';
-import updateProjectSource from './updateProjectSource';
+import errors from './errors';
 
 export default [
   linkGithubIdentity,
   logout,
   startAccountMigration,
   unlinkGithubIdentity,
-  updateProjectSource,
+  ...errors,
 ];
