+++ src/sagas/index.js
@@ -1,7 +1,6 @@
 import {all} from 'redux-saga/effects';
 
 import manageUserState from './manageUserState';
-import watchErrors from './errors';
 import watchProjects from './projects';
 import watchUi from './ui';
 import watchClients from './clients';
@@ -11,7 +10,6 @@ import watchAssignments from './assignments';
 export default function* rootSaga() {
   yield all([
     manageUserState(),
-    watchErrors(),
     watchProjects(),
     watchUi(),
     watchClients(),
